finished front-end

+++ client/src/admin/Dashboard.jsx
@@ -8,6 +8,11 @@ const API_URL = import.meta.env.VITE_API_URL;
 const Dashboard = () => {
   OngletTitle("Dashboard Admin");
   const [users, setUsers] = useState([]);
+  const [showModal, setShowModal] = useState(false);
+
+  const handleShowModal = () => setShowModal(true);
+
+  const handleCloseModal = () => setShowModal(false);
 
   useEffect(() => {
     const fetchUsers = async () => {
@@ -40,7 +45,7 @@ const Dashboard = () => {
         {users.map((user) => (
           <li
             key={user._id}
-            className="list-group-item d-flex justify-content-between align-items-center"
+            className="list-group-item d-flex justify-content-between align-items-center w-75 mx-auto"
           >
             <div>
               <p className="mb-1">
@@ -50,11 +55,52 @@ const Dashboard = () => {
             <div>
               <button
                 className="btn btn-danger w-100"
-                onClick={() => deleteUser(user._id)}
+                onClick={handleShowModal}
               >
                 Supprimer
               </button>
             </div>
+            {showModal && (
+              <div
+                className="modal"
+                style={{
+                  display: "block",
+                  backgroundColor: "rgba(0, 0, 0, 0.5)",
+                }}
+              >
+                <div className="modal-dialog modal-dialog-centered">
+                  <div className="modal-content">
+                    <div className="modal-header">
+                      <h5 className="modal-title">Confirmer la suppression</h5>
+                      <button
+                        type="button"
+                        className="btn-close"
+                        onClick={handleCloseModal}
+                      ></button>
+                    </div>
+                    <div className="modal-body">
+                      Etes-vous sur de vouloir supprimer cet utilisateur ?
+                    </div>
+                    <div className="modal-footer">
+                      <button
+                        type="button"
+                        className="btn btn-secondary"
+                        onClick={handleCloseModal}
+                      >
+                        Annuler
+                      </button>
+                      <button
+                        type="button"
+                        className="btn btn-danger"
+                        onClick={() => deleteUser(user._id)}
+                      >
+                        Supprimer
+                      </button>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            )}
           </li>
         ))}
       </ul>
